fixed training network
Result: The training of the networks works, but the test of the algorithm shows, that the applied linear regression model is not that accurate for this problem.
After defining the weights of the network, the input frequencies where multiplied to the weights and they have significant derivations to the original ones in the dataset. Thats why another network is needed to increase the accuracy.
On other hand the both weights are quite similar, thats why the datasets of both frequencies with both photosynthetic acitivities get calculated together in future.